### 2024-03-14 e07b4df
Add DOI to Skyrms book in paper.bib
