+++ docs/paper/paper.bib
@@ -21,6 +21,7 @@
   publisher = {{Oxford University Press}},
   location = {{Oxford}},
   isbn = {978-0-19-958082-8},
+  doi = {10.1093/acprof:oso/9780199580828.001.0001},
   langid = {english},
   pagetotal = {208},
   keywords = {Philosophy / Epistemology,Philosophy / Language,Philosophy / Mind \& Body,Science / Life Sciences / Evolution,Science / Philosophy \& Social Aspects}
