### 2024-04-30 #6703 @editorialbot
[REVIEW]: evoke: A Python package for evolutionary signalling games
**Submitting author:** @stephenfmann (<a href="http://orcid.org/0000-0002-4136-8595">Stephen Francis Mann</a>)
**Repository:** https://github.com/signalling-games-org/evoke
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
