### 2024-06-12 #6703 @Nikoleta-v3
Apologies for the delay in my review; the past few weeks have been very busy. Anyway, I had some time to go over the submission now, and I have a few comments. I have some minor comments, but also a major concern about the package. This concern regards the documentation of the package and its overall aim.

Reading through the documentation and the paper, it feels that the aim of the package is to replicate two pieces of work: Skyrms 2010 and Martinez 2013. This is a very limited usage of a package.

In the current documentation, there is no explanation of what signaling games are, and instructions on how to create your own signaling games appear only in the later section of the interactive tutorial. Let’s assume that the target audience consists of people familiar with signaling games and the evolutionary algorithms you are using. The first things that show in the tutorials and documentation are how to replicate the two prior works. However, even here there is no explanation given; the functions are even named after the figure numbers. For example, `Skyrms2010_5_2()` replicates figure 5.2. What if I have no idea what figure 5.2 of the paper is?

The package should serve as a tool for creating signaling games and running evolutionary simulations. The secondary aim of the package would be replicating current pieces of work. I am not saying that these should not be mentioned or part of the documentation; please see my suggestions in the Documentation section.

Let me break down my comments in the following sections:

**Paper**

I find the paper lacking because of what I have already mentioned. The scope of the package is too narrow. In the summary, it is even stated that “More examples from well-known books and papers will be added”

The package should be about what your second sentence says: creating both reconstructions of models from the literature and their own models and results. It’s just that there is no documentation on how to do this.

There are several packages on game theory, for example, see: https://egttools.readthedocs.io/en/latest/tutorials.html. However, there are no packages (to my knowledge) on signaling games. So do I believe there is a need for such a package.
That being said maybe consider citing some of there other packages? This is another well know Python package for example: https://nashpy.readthedocs.io/en/stable/#.

### Documentation

Currently, there are the online material, the online document, and the examples folder, and all cover different things. I like the interactive documentation, but these sections should also be in the online documentation.

The online document should have sections that describe signaling games and how to create them, the description of the processes, and how to run them. Then follow with the plotting. Replicating prior literature should have their own sections, that is okay, but you also need to describe what those papers were doing, not just say here you can get figure 1.1.

I know that in the evoke/examples scripts, the docstrings of the classes have the information, and they also appear in the examples: https://evoke.readthedocs.io/en/latest/examples.html. But it’s information that the user has to go look for.

I suggest that you focus on the Sphinx online document and then also have the interactive material that covers some functionality of the package.

### Minor

- Python versions: The package requires Python greater than 3.8. That is documented on PyPI; maybe include something on the repository too? Are you familiar with badges?
- Question: I see you have a license file, but it is called COPYING. How come? It’s the first time I've seen this name; it’s usually called LICENSE.
- Please add a contributions file. In the interactive material, you have “Pull request: When you've tested and everything looks good, create a pull request from your fork.” However, this cannot be found anywhere else, like in the repository or the online documentation. Please include your contribution line in the README as well. Also, please consider having more detailed guidelines. See examples: https://gitlab.com/remram44/taguette/-/blob/master/CONTRIBUTING.md?ref_type=heads.
- Could you also include some details on how to run your tests? If you choose to have more detailed guidelines for contributing, as you can see in the example, this running tests section is commonly included there.

### Tests

Regarding your tests, I managed to run all of them and they passed. However, what are you testing? All the tests are just testing whether something is an instance. For example, consider the class `OnePop` you could test that all the input arguments are passed correctly.

However, since tests are not required for a JOSS publication, this is a recommendation.

I am requesting a lot of work. I believe that such a package is missing from the literature, but I find the package to be a bit weak in its current form.
