### 2024-06-12 #6703 @stephenfmann
Thanks, that all sounds sensible. We were assuming quite a bit of background knowledge, and we will update the documentation and the paper so as not to assume so much.

Re: testing, this is something I don't have a lot of experience with, and so it's not clear to me what exactly to test other than that the object was created correctly (hence the checking for instances). When you say "you could test that all the input arguments are passed correctly", do you mean to create the object and then check that the properties of that object are equal to the values that were passed into it? I don't really have an intuition as to what makes a good test, but maybe I can read up on that.

Re: the figures, I'm concerned about copyright issues, so I don't want to actually depict the figure from the book. This is annoying, but it's a consequence of the fact that the Skyrms book is the classic text in this area, and so we chose to focus on it for the main example.

When you say: "The package should be about what your second sentence says: creating both reconstructions of models from the literature and their own models and results. It’s just that there is no documentation on how to do this." It's in the interactive tutorial, but I guess you mean it should be in the ReadTheDocs pages?

When you say: "Currently, there are the online material, the online document, and the examples folder, and all cover different things." The examples folder has to be more extensive than the other two, because it's where examples from the literature go. For the other two, if by "online material" you mean the interactive tutorial, it covers the basics whereas the documentation is supposed to cover everything in detail. Or does "the online document" mean the interactive tutorial? I'm a bit confused.

Anyway, I plan to action all your comments, so thank you very much for your efforts!
