### 2024-07-19 1711c6a
Update README: add details of example figures
