### 2024-07-22 3b743a8
JOSS paper: Add references to Nashpy and EGTtools
