+++ docs/paper/paper.bib
@@ -26,3 +26,21 @@
   pagetotal = {208},
   keywords = {Philosophy / Epistemology,Philosophy / Language,Philosophy / Mind \& Body,Science / Life Sciences / Evolution,Science / Philosophy \& Social Aspects}
 }
+
+@misc{Fernandez2020,
+  author = {Fernández Domingos, Elias},
+  title = {EGTTools: Toolbox for Evolutionary Game Theory},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/Socrats/EGTTools}},
+  doi = {10.5281/zenodo.3687125}
+}
+
+@misc{nashpyproject,
+  author       = {{ {The Nashpy project developers} }},
+  title        = {Nashpy: 0.0.41},
+  year         = {2024},
+  doi          = {10.5281/zenodo.10802174},
+  url          = {http://dx.doi.org/10.5281/zenodo.10802174}
+}
