### 2024-07-22 37e7ba2
Add python version requirement to README
