### 2024-07-22 a219f9e
Created CONTRIBUTING.md
