### 2024-07-22 a594289
Update CONTRIBUTIONS.md and add contrib info to README.md
