### 2024-07-22 0b09bdc
README: the correct filename is CONTRIBUTING.md
