### 2024-07-25 a827a06
Add test information to CONTRIBUTING.md
