### 2024-07-26 8bcdc84
Capitalise Evoke everywhere
