+++ evoke/examples/skyrms2010signals.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """
 
-evoke library examples from:
+Evoke library examples from:
     
     Skyrms, B. (2010). Signals: Evolution, Learning, and Information. Oxford University Press.
 """
