+++ docs/source/index.rst
@@ -1,4 +1,4 @@
-.. evoke documentation master file, created by
+.. Evoke documentation master file, created by
    sphinx-quickstart on Fri Aug 18 12:31:09 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
@@ -6,7 +6,7 @@
 Evoke: Evolutionary signalling games with Python
 ================================================
 
-**evoke** is a Python library for evolutionary simulations of signalling games.
+**Evoke** is a Python library for evolutionary simulations of signalling games.
 It is particularly oriented towards reproducing results and figures from the literature, and offers a simple and intuitive API.
 
 * `Tutorial <https://colab.research.google.com/drive/1AwUCP05lpITAP7_EZD7loGv3unhnwvhM#forceEdit=true&sandboxMode=true>`_ | *Google Colab*
