### 2024-07-26 c9108b4
Consistent title for README and documentation
