+++ docs/source/usage.rst
@@ -18,30 +18,10 @@ Tutorial
 The best way to discover Evoke is `the interactive tutorial <https://colab.research.google.com/drive/1AwUCP05lpITAP7_EZD7loGv3unhnwvhM#forceEdit=true&sandboxMode=true>`_.
 The rest of this document describes very simple use cases.
 
+Creating signalling games
+-------------------------
 
-The simplest case: reproducing a figure from the literature
------------------------------------------------------------
-
-The easiest thing to do with Evoke is recreate a figure from the signalling game literature.
-The parameters required to create some of these figures are included in the ``examples`` folder.
-
-For example, to create figure 5.2 from *Signals* (Skyrms 2010), you would run:
-
-.. code-block:: python
-
-   Skyrms2010_5_2()
-
-This creates an **object** which runs an evolutionary simulation with parameters as close as possible to those described by Bryan Skyrms for the figure in the book.
-
-In this case Evoke creates a figure very close to the original.
-In other cases there might be a range of random properties that can cause deviation from the figures in the literature.
-It's often a good idea to create the same figure multiple times, to get a feel for the range of variation that can be produced by the reported parameters.
-
-
-Creating simulations
---------------------
-
-If you want to create custom simulations, you need to create two kinds of object:
+Creating signalling game simulations requires two kinds of object:
 
 - **game**: This describes properties of the game, including the probabilities of each observable state, the number of signals available, and the payoff matrices of sender and receiver.
 - **evolve**: This describes properties of the evolutionary scenario, especially whether the 'agents' are populations evolving via selection or individuals learning via reinforcement.
@@ -49,3 +29,180 @@ If you want to create custom simulations, you need to create two kinds of object
 Games and evolve objects can be mixed and matched.
 This allows you to see differences between evolution and reinforcement learning, by taking the same game and plugging it into different evolve objects.
 
+Take the simple example of a cooperative game with three states, three signals and three acts (the following code examples can also be found in the `tutorial <https://colab.research.google.com/drive/1AwUCP05lpITAP7_EZD7loGv3unhnwvhM#forceEdit=true&sandboxMode=true&scrollTo=304dd76f>`_).
+First, in order to create the **game** object, we need to supply Evoke with four pieces of information:
+
+- `state_chances`: The number of states the sender can observe, and their probabilities. For this example there will be three equiprobable states.
+- `sender_payoff_matrix` and `receiver_payoff_matrix`: Matrices defining the payoffs of sender and receiver. For this example they will be the same: when the state matches the act, both agents get a payoff of 1, otherwise they get a payoff of zero.
+- `messages`: The number of messages available for the sender to send, which in this example is three.
+
+.. code-block:: python
+
+    state_chances = np.array([1/3, 1/3, 1/3])
+    sender_payoff_matrix = np.eye(3)
+    receiver_payoff_matrix = np.eye(3)
+    messages = 3
+
+We import the `Chance` object from the `game` module, because this kind of game starts with a state of the world that is determined by chance.
+Our game will be an instance of this object:
+
+.. code-block:: python
+
+    from evoke.src.games import Chance
+    
+    game = Chance(
+        state_chances   = state_chances,
+        sender_payoff_matrix = sender_payoff_matrix,
+        receiver_payoff_matrix = receiver_payoff_matrix,
+        messages = messages
+    )
+
+Second, to create the **evolve** object, we need to specify which strategies the agents will be using at the beginning of the simulation.
+Fortunately, the game object already knows all the strategies that are in principle available to an agent:
+
+.. code-block:: python
+
+    sender_strategies = game.sender_pure_strats()
+    receiver_strategies = game.receiver_pure_strats()
+
[…]
