### 2024-07-29 1c54321
Allow python 3.12; update requirements text
