### 2024-07-29 da3e900
pygambit strictly needs to be version 16.1.0 or 16.1.1
