### 2024-07-29 fd95d3d
Fix time vector bug and MatchingSIR stats bug
