### 2024-07-29 e6fd29b
More thorough evolve tests; suppress figure creation on test; pygambit requirement in test
