+++ evoke/tests/test_evolve.py
@@ -71,15 +71,20 @@ class TestEvolve(unittest.TestCase):
 
         # Just check these don't crash
         evo.avg_payoff_vector(np.array([1, 0, 0, 0]))
-        evo.discrete_replicator_delta_X(np.array([1 / 4, 1 / 4, 1 / 4, 1 / 4]))
         evo.pop_to_mixed_strat(np.array([1 / 4, 1 / 4, 1 / 4, 1 / 4]))
         evo.random_player()
-        evo.replicator_dX_dt_odeint(np.array([0, 1 / 3, 1 / 3, 1 / 3]), 0)
-        evo.replicator_discrete(np.array([1 / 8, 1 / 4, 1 / 4, 1 / 8]), 4)
         evo.replicator_jacobian_odeint(np.array([0, 1 / 3, 1 / 3, 1 / 3]))
-        evo.replicator_odeint(
+
+        # Check the replicator equations yield vectors that sum to 1
+        self.assertEqual(1,evo.discrete_replicator_delta_X(np.array([1 / 4, 1 / 4, 1 / 4, 1 / 4])).sum()) # 1 step
+        self.assertEqual(1,evo.replicator_discrete(np.array([1 / 8, 1 / 4, 1 / 4, 1 / 8]), 4)[-1].sum()) # 4 steps
+        self.assertEqual(1,evo.replicator_odeint(
             np.array([6 / 16, 1 / 16, 4 / 16, 5 / 16]), range(10)
-        )  # 10 time steps
+        )[-1].sum()) # 10 steps
+
+        # Check the continuous replicator dX/dt vector sums to 0 (or nearly so)
+        self.assertAlmostEqual(0,evo.replicator_dX_dt_odeint(np.array([0, 1 / 3, 1 / 3, 1 / 3]), 0).sum(),5)
+        
 
         # Check is instance
         self.assertIsInstance(evo, evolve.OnePop)
@@ -122,16 +127,23 @@ class TestEvolve(unittest.TestCase):
         # This method expects a 1D array that just concatenates the sender pop
         # and the receiver pop.
 
-        evo.discrete_replicator_delta_X(random_population)
+        # Just check these don't crash
         evo.receiver_avg_payoff(random_sender, random_receiver)
         evo.receiver_to_mixed_strat(random_receiver)
-        evo.replicator_dX_dt_ode(t=0, X=random_population)
-        evo.replicator_dX_dt_odeint(X=random_population, t=0)
-        evo.replicator_jacobian_ode(t=0, X=random_population)
-        evo.replicator_jacobian_odeint(random_population)
         evo.sender_avg_payoff(random_sender, random_receiver)
         evo.sender_to_mixed_strat(random_sender)
-        evo.vector_to_populations(random_population)
+        evo.replicator_jacobian_ode(t=0, X=random_population)
+        evo.replicator_jacobian_odeint(random_population)
+
+        # Check vector_to_populations yields an array of length 2
+        self.assertEqual(2,len(evo.vector_to_populations(random_population)))
+
+        # Check the replicator equations yield vectors that sum to 2 (that's 1 per population)
+        self.assertAlmostEqual(2,evo.discrete_replicator_delta_X(random_population).sum(),5) # 1 step
+
+        # Check the continuous replicator dX/dt vector sums to 0 (or nearly so)
+        self.assertAlmostEqual(0,evo.replicator_dX_dt_ode(t=0, X=random_population).sum(),5)
+        self.assertAlmostEqual(0,evo.replicator_dX_dt_odeint(X=random_population, t=0).sum(),5)
 
         # Create evolve.Times object to test replicator methods
         initial_time = 0
@@ -139,10 +151,10 @@ class TestEvolve(unittest.TestCase):
[…]
