### 2024-07-30 86c2959
Update documentation: code snippets in index
