+++ docs/source/index.rst
@@ -7,7 +7,7 @@ Evoke: A Python package for evolutionary signalling games
 =========================================================
 
 **Evoke** is a Python library for evolutionary simulations of signalling games.
-It is particularly oriented towards reproducing results and figures from the literature, and offers a simple and intuitive API.
+It offers a simple and intuitive API that can be used to analyze arbitrary game-theoretic models, and to easily reproduce and customize well-known results and figures from the literature.
 
 * `Tutorial <https://colab.research.google.com/drive/1AwUCP05lpITAP7_EZD7loGv3unhnwvhM#forceEdit=true&sandboxMode=true>`_ | *Google Colab*
 * `Documentation <https://evoke.readthedocs.io/en/latest/>`_ **You are here!** | *ReadTheDocs*
@@ -40,12 +40,19 @@ The following section shows how each of these components can be defined for a si
 A simple signalling game
 ------------------------
 
+We will describe one of the simplest possible signalling games, along with code snippets illustrating how Evoke represents each game element.
+A fuller demonstration can be found in the section :ref:`Creating signalling games`.
+
 Consider two agents, a sender and a receiver, who need to cooperate in order to achieve a common goal.
-On each round of the game, one out of two possible states of the world obtains with 50/50 probability.
+In each round of the game, one out of two possible states of the world obtains with 50/50 probability.
+
+.. code-block:: python
+
+   state_chances = np.array([1/2, 1/2])
+
 The sender observes which state occurs, and sends one of two possible signals to the receiver.
 Upon receipt of the signal, the receiver performs one of two possible acts.
-Payoffs to sender and receiver are identical and are determined by whether or not the receiver's act matches the state observed by the sender.
-Payoffs can be represented in a table:
+Payoffs to sender and receiver are identical and are determined by whether or not the receiver's act matches the state observed by the sender:
 
 ============= =======  =======
  *Payoffs*     Act 1    Act 2 
@@ -54,31 +61,89 @@ Payoffs can be represented in a table:
  **State 2**     0        1   
 ============= =======  =======
 
+Evoke represents these payoffs as a 2x2 matrix:
+
+.. code-block:: python
+
+   sender_payoff_matrix = receiver_payoff_matrix = np.array([[1, 0], [0, 1]])
+
+With these elements in place, we can create a signalling game object:
+
+.. code-block:: python
+
+   from evoke.src.games import Chance
+
+   # Chance game: the state of the world is determined probabilistically
+   game = Chance( state_chances, 
+                  sender_payoff_matrix, 
+                  receiver_payoff_matrix,
+                  messages = 2
+               )
+
 We have two agents, the basic actions available to them, and the payoffs associated with the outcomes that result from the combination of these actions.
+
 What about strategies?
 An agent's strategy describes which actions it performs in which circumstance.
 In this simple game the sender has four deterministic or "pure" strategies, corresponding to the different ways it can produce signals in response to the states of the world it observes:
 
+- State 1 => Signal 1; State 2 => Signal 1
 - State 1 => Signal 1; State 2 => Signal 2
 - State 1 => Signal 2; State 2 => Signal 1
-- State 1 => Signal 1; State 2 => Signal 1
 - State 1 => Signal 2; State 2 => Signal 2
 
-The latter two strategies are effectively useless to the receiver, since they entail sending the same signal no matter what the state of the world.
+Each strategy is represented as a matrix.
+The matrix has two rows, corresponding to the two states of the world, and two columns, corresponding to the two possible signals.
[…]
