+++ docs/source/conf.py
@@ -29,8 +29,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx.ext.intersphinx',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.autosectionlabel',
+    'sphinx.ext.napoleon'
 ]
 
 intersphinx_mapping = {
