### 2024-07-30 #6703 @stephenfmann
Thanks @Nikoleta-v3 for your patience. We have now addressed all of the issues discussed above. I'll list the changes in response to the original review.

> In the current documentation, there is no explanation of what signaling games are, and instructions on how to create your own signaling games appear only in the later section of the interactive tutorial. Let’s assume that the target audience consists of people familiar with signaling games and the evolutionary algorithms you are using. The first things that show in the tutorials and documentation are how to replicate the two prior works. However, even here there is no explanation given; the functions are even named after the figure numbers. For example, Skyrms2010_5_2() replicates figure 5.2. What if I have no idea what figure 5.2 of the paper is?

The [documentation](https://evoke.readthedocs.io/en/latest/) has been heavily revised. The **Introduction** describes both games in general and signalling games in particular, and details the components of a signalling game along with code snippets. The **Usage** page gives a full demonstration of creating and running a signalling game simulation, with code snippets taken from the interactive tutorial. Each time a figure appears, its meaning is explained so that the reader understands what they are looking at.

> The package should serve as a tool for creating signaling games and running evolutionary simulations. The secondary aim of the package would be replicating current pieces of work. I am not saying that these should not be mentioned or part of the documentation; please see my suggestions in the Documentation section.

The emphasis of the documentation, paper, tutorial and README are now on creating signalling games and running evolutionary simulations. In each of those four documents, information about creating and simulating signalling games comes before details of replicating existing figures.

> I find the paper lacking because of what I have already mentioned. The scope of the package is too narrow. In the summary, it is even stated that “More examples from well-known books and papers will be added”
> The package should be about what your second sentence says: creating both reconstructions of models from the literature and their own models and results. It’s just that there is no documentation on how to do this.

The paper now emphasises creating and simulating a user's own games, while still mentioning the possibility of recreating figures from the literature. As mentioned above, the documentation now goes into detail about how to create and simulate custom games.

> There are several packages on game theory, for example, see: https://egttools.readthedocs.io/en/latest/tutorials.html. However, there are no packages (to my knowledge) on signaling games. So do I believe there is a need for such a package.
That being said maybe consider citing some of there other packages? This is another well know Python package for example: https://nashpy.readthedocs.io/en/stable/#.

The paper now references both Nashpy and EGTtools, while emphasising that the focus on signalling games is what is novel about this package.

> Currently, there are the online material, the online document, and the examples folder, and all cover different things. I like the interactive documentation, but these sections should also be in the online documentation.

All of the key information from the tutorial now appears in the documentation too.

> The online document should have sections that describe signaling games and how to create them, the description of the processes, and how to run them. Then follow with the plotting. Replicating prior literature should have their own sections, that is okay, but you also need to describe what those papers were doing, not just say here you can get figure 1.1.

The documentation now extensively describes signalling games and how they are represented in Evoke. The figure example is described in detail, and a link to the relevant docstring is included.

> I know that in the evoke/examples scripts, the docstrings of the classes have the information, and they also appear in the examples: https://evoke.readthedocs.io/en/latest/examples.html. But it’s information that the user has to go look for.

As mentioned, the docstring of the figure in the documentation is directly linked. The figures in the tutorial and README are now explained in detail (i.e. the information in the docstring is reproduced, though not necessarily verbatim).

> I suggest that you focus on the Sphinx online document and then also have the interactive material that covers some functionality of the package.

We have followed this advice in the new version. The Sphinx documentation (hosted at [ReadTheDocs](https://evoke.readthedocs.io/en/latest/index.html)) is now much more extensive.

> Python versions: The package requires Python greater than 3.8. That is documented on PyPI; maybe include something on the repository too? Are you familiar with badges?

We have added the Python version requirement to the text of the README along with a link to `pyproject.toml` which lists other requirements. We have also added a badge to the top of the README. The next release will be compatible with Python 3.12; because the badge is linked to PyPI, it will update only when the new release is uploaded, which we plan to do after the review phase is completed.

> Question: I see you have a license file, but it is called COPYING. How come? It’s the first time I've seen this name; it’s usually called LICENSE.

We have changed the filename to LICENSE. The reason it was COPYING was that that is the name [recommended by the Free Software Foundation](https://www.gnu.org/licenses/gpl-howto.html).

> Please add a contributions file. In the interactive material, you have “Pull request: When you've tested and everything looks good, create a pull request from your fork.” However, this cannot be found anywhere else, like in the repository or the online documentation. Please include your contribution line in the README as well. Also, please consider having more detailed guidelines. See examples: https://gitlab.com/remram44/taguette/-/blob/master/CONTRIBUTING.md?ref_type=heads.

We have added a [CONTRIBUTING.md](https://github.com/signalling-games-org/evoke/blob/main/CONTRIBUTING.md) (modelled on the linked file) and we link to it in the README.

> Could you also include some details on how to run your tests? If you choose to have more detailed guidelines for contributing, as you can see in the example, this running tests section is commonly included there.

In CONTRIBUTING.md we have added a section on running tests.

> Regarding your tests, I managed to run all of them and they passed. However, what are you testing? All the tests are just testing whether something is an instance. For example, consider the class OnePop you could test that all the input arguments are passed correctly.
> However, since tests are not required for a JOSS publication, this is a recommendation.

We have added much more detail to `test_evolve.py` and plan to update the other tests too (as you say, this is not a strict requirement so we hope it isn't a showstopper). In fact, in updating the tests we found two minor bugs, so thank you!

Thanks so much for your careful review. We hope we have addressed all the issues; if not we will take another look at anything that is still amiss.
