+++ MANIFEST.in
@@ -1,4 +1,0 @@
-include COPYING
-include README.md
-include setup.py
-recursive-include skyrms *
