+++ pyproject.toml
@@ -3,7 +3,7 @@ name = "evoke-signals"
 packages = [
     { include="evoke", from="." },
 ]
-version = "0.1.3"
+version = "0.1.4"
 description = "Signalling games in python"
 authors = [
 		"Stephen Mann <stephenfmann@gmail.com>", 
