### 2024-08-12 #6703 @Nikoleta-v3
Hello @stephenfmann 👋🏻

Thank you very much for addressing my comments! Both the documentation and the paper have improved significantly.

> We have added the Python version requirement to the README along with a link to `pyproject.toml`, which lists other requirements. We have also added a badge to the top of the README. The next release will be compatible with Python 3.12; because the badge is linked to PyPI, it will update only when the new release is uploaded, which we plan to do after the review phase is completed.

Amazing! and the badge looks great on the README 😄

> We have changed the filename to LICENSE. The reason it was previously named COPYING is that this is the name recommended by the Free Software Foundation.

Interesting! I hadn’t seen that before.

> We have added a [CONTRIBUTING.md](https://github.com/signalling-games-org/evoke/blob/main/CONTRIBUTING.md) (modeled on the linked file) and we link to it in the README.

The contributing file looks great, especially the "How to Run Tests" section! 👍🏻

> We have added much more detail to `test_evolve.py` and plan to update the other tests too (as you mentioned, this is not a strict requirement, so we hope it isn't a showstopper). In fact, while updating the tests, we found two minor bugs, so thank you!

Haha, I’m glad it was helpful!

Regarding the paper I have 2️⃣  last minor comments. Apart from that @drvinceknight I’m happy to approve the paper for publication! 🚢

Comments:

(-) Line 17 should be `game dynamics`?
(-) Line 21-23.  It sounds like there could have been more examples from books included, but they weren't. **Maybe** you could rephrase it to say that users can easily contribute to the library by adding examples from literature, making it a useful educational tool. At the same time, they can create and analyze their own models, turning it into a great research resource.
