### 2024-08-12 #6703 @stephenfmann
Thanks so much @Nikoleta-v3 ! For the sentence "Evoke offers methods for both kinds of game dynamic", here the singular "game dynamic" I think is correct. We are saying that reinforcement learning is one dynamic and evolution is a different dynamic. It's a bit weird but people do sometimes use "dynamic" in the singular like this. (However, if you happen to know that vastly many more game theorists would use "dynamics" in this context, I'll be happy to change it.)

For lines 21-23 I have rephrased these sentences. The text now reads: "Users can contribute to the library by adding further examples from the literature. This can be a useful way to become familiar with Evoke, while at the same time increasing the benefit to other users. Evoke can therefore serve as an educational tool (encouraging understanding of existing literature) and a research resource (promoting good practice and effective modelling techniques)."

The new paper is [here](https://github.com/openjournals/joss-papers/blob/joss.06703/joss.06703/10.21105.joss.06703.pdf). Hope this is satisfactory!
