### 2024-09-21 #6703 @stephenfmann
Hi @danielskatz @drvinceknight how can we help you push this forward? Other potential reviewers from the JOSS list might be @marimeireles, @mwt and @igarizio.
