### 2024-10-24 signalling-games-org/evoke#72 @igarizio
Broken colab tutorial link?
Hi @stephenfmann,

I think the [colab tutorial link](https://colab.research.google.com/drive/1AwUCP05lpITAP7_EZD7loGv3unhnwvhM#forceEdit=true&sandboxMode=true) in the README, docs, etc is broken (or at least I can't access it).
![drive](https://github.com/user-attachments/assets/2a84e82f-229d-4564-872b-443bc210a584)

Thanks!
