### 2024-10-24 signalling-games-org/evoke#72 @stephenfmann
I've restored the notebook but it will now have a different URL. You should be able to access it [here](https://colab.research.google.com/drive/1zFRb20KoQi0Tdg5UR1_zReHCZicnQLZd#forceEdit=true&sandboxMode=true). I will update the URL in the various places where it's listed, and then close the issue.

Thanks for this! Let me know if there are any more problems.
