### 2024-10-24 signalling-games-org/evoke#72 @stephenfmann
Links updated in README and documentation. The link on the PyPI package page will be updated next time there's a version release, which is planned for after the JOSS review process.
