### 2024-10-25 signalling-games-org/evoke#72 @igarizio
Hi, sorry to comment again, but now there is a permission issue ("You need access... Request access, or switch to an account with access.").

My apologies for not suggesting this earlier, but I think it would be a good idea to have your notebook saved in the repository (as an `.ipynb` file). This will not only resolve any access issues but will let us have an archived copy of it.
Once you have the file, you can update all the colab links like this (and everything should work the same):
- Example notebook: `https://github.com/jupyter/notebook/blob/main/docs/source/examples/Notebook/Running%20Code.ipynb`
- Colab link: `https://colab.research.google.com/github/jupyter/notebook/blob/main/docs/source/examples/Notebook/Running%20Code.ipynb`

In other words, you put `https://colab.research.google.com/` + github link removing `https://` and `.com`.

Please let me know what you think. Thanks again!
