### 2024-10-25 signalling-games-org/evoke#72 @stephenfmann
Thanks, the access issue should be fixed now. Could you check again?

I'll have a look at embedding the notebook in the repository. The last time I tried there were some frustrating issues, e.g. git wanted to sync all of the outputs of the cells after I had finished browsing the notebook. But maybe there are ways around that kind of behaviour.
