### 2024-10-25 signalling-games-org/evoke#72 @stephenfmann
Thanks, I've opened a new issue to investigate moving the notebook into the repository. I'll close this issue now (but feel free to reopen it if you have any more issues!)
