### 2024-11-10 #6703 @drvinceknight
Hi @igarizio how are things going with this?
