### 2024-11-19 #6703 @igarizio
Hi everyone, quick update, I have finished the review and everything looks good.
For reference, here are some issues I created:
- [Overall structure and style](https://github.com/signalling-games-org/evoke/issues/75)
- [Irrelevant testing](https://github.com/signalling-games-org/evoke/issues/76)
- [Broken colab tutorial link?](https://github.com/signalling-games-org/evoke/issues/72)

@stephenfmann addressed all these and I have now checked every checkbox in the review.
