### 2024-11-19 #6703 @stephenfmann
The repository is at Zenodo with DOI [10.5281/zenodo.14185732](https://doi.org/10.5281/zenodo.14185732). I set the version number as 0.1.4, because a few (minor) changes have been made since the last PyPI release at 0.1.3. I hope this is OK.
