### 2024-11-20 #6703 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
