### 2024-11-20 #6703 @danielskatz
Congratulations to  @stephenfmann (Stephen Francis Mann) and co-author on your publication!!

JOSS depends on volunteers and we couldn't do this work without you
